OceanRasterConversions.jl documentation

Overview

This package converts ocean varaibles that are saved as Raster data structures using GibbsSeaWater.jl. Rasters.jl provides excellent reading, writing and manipulation of geospatial data. Typically, the salt and temperature variables from ocean models or observational data are practical salinity and potential temperature so conversions must be to the TEOS-10 standard variables of absolute salinity and conservative temperature to accurately calculate further variables like seawater density.

Further conversions and other water mass transformation procedures will be added in the future.

Package workings

This package will convert the variables practical salinity and potential temperature into absolute salinity and conservative temperature. In doing so a pressure variable is needed, so this is created and returned in the RasterStack. Lastly a density variable (either in-situ or potential referenced to a user input) is computed and added to the RasterStack. See the example for how the package can be used.

Variables

The variables are named using the symbols that represent them. The symbols are unicode characters which can be generated in the julia repl by pressing tab after the varible

julia> \theta#press tab

will autocomplete to θ, the symbol for potential temperature. The subscript letters that are used to distinguish between practical salinity, Sₚ, and absolute salinity, Sₐ, are also added in the julia repl

julia> S\_a#press tab

Currently the varabile symbols are:

θ potential temperature

Θ conservative temperature

Sₚ practical salinity

Sₐ absolute salinity

p pressure

ρ in-situ seawater density

σₚ potential density at user defined reference pressure ₚ.

Limitations

If the required dimensions for the conversions are not present an error will be thrown. For example, trying to convert a RasterStack that has no depth dimension will not work as the Z dimension is not found and the pressure variable depends on depth. There is a manual workaround for this. When defining the RasterStack add the Z dimension as a single entry, rather than a Vector,

lons, lats, z = -180:180, -90:90, 0.0
stack = RasterStack(data, (X(lons), Y(lats), Z(z)))

This is equivalent to a two dimensional RasterStack at sea-surface height (z = 0).

Currently the only dimension names that are supported are X, Y, Z, and Ti. Allowing for user specified dimensions has not yet been implemented.

Functions exported from OceanRasterConversions

OceanRasterConversions.convert_ocean_vars — Function

function convert_ocean_vars(raster::RasterStack, var_names::NamedTuple;
                            ref_pressure = nothing,
                            with_α = false,
                            with_β = false)
function convert_ocean_vars(raster::Rasterseries, var_names::NamedTuple;
                            ref_pressure = nothing,
                            with_α = false,
                            with_β = false)

Convert ocean variables depth, practical salinity and potential temperature to pressure, absolute salinity and conservative temperature. All conversions are done using the julia implementation of TEOS-10 GibbsSeaWater.jl. A new Raster is returned that contains the variables pressure, absolute salinity, conservative temperature and density (either in-situ or referenced to a user defined reference pressure). As pressure depends on latitude and depth, it is added as a new variable –- that is, each longitude, latitude, depth and time have a variable for pressure. A density variable is also computed which, by default, is in-situ density. Potential density at a reference pressure can be computed instead by passing a the keyword argument ref_pressure. Optional keyword arguments with_α and with_β allow the thermal expansion and haline contraction coefficients (respectively) to be computed and added to the returned RasterStack/Series.

The name of the variables for potential temperature and practical salinity must be passed in as a NamedTuple of the form (Sₚ = :salt_name, θ = :potential_temp_name) where :potential_temp_name and :salt_name are the name of the potential temperature and practical salinity in the Raster.

OceanRasterConversions.depth_to_pressure — Function

function depth_to_pressure(raster::Raster)
function depth_to_pressure(stack::RasterStack)

Convert the depth dimension (Z) to pressure using gsw_p_from_z from GibbsSeaWater.jl. Note that pressure depends on depth and latitude so the returned pressure is stored as a variable in the resulting Raster rather than replacing the vertical depth dimension.

OceanRasterConversions.Sₚ_to_Sₐ — Function

function Sₚ_to_Sₐ(Sₚ::Raster)
function Sₚ_to_Sₐ(stack::RasterStack, Sₚ::Symbol)
function Sₚ_to_Sₐ(series::RasterSeries, Sₚ::Symbol)

Convert a Raster of practical salinity (Sₚ) to absolute salinity (Sₐ) using gsw_sa_from_sp from GibbsSeaWater.jl. This conversion depends on pressure. If converting from a RasterStack or RasterSeries, the symbol for the practical salinity in the RasterStack/Series must be passed in as a Symbol –- that is if the variable name is SALT the RasterStack/Series, the Symbol :SALT must be passed in.

OceanRasterConversions.θ_to_Θ — Function

function θ_to_Θ(θ::Raster, Sₐ::Raster)
function θ_to_Θ(stack::RasterStack, var_names::NamedTuple)
function θ_to_Θ(series::RasterSeries, var_names::NamedTuple)

Convert a Raster of potential temperature (θ) to conservative temperature (Θ) using gsw_ct_from_pt from GibbsSeaWater.jl. This conversion depends on absolute salinity. If converting from a from a RasterStack or a RasterSeries, the var_names must be passed in as for convert_ocean_vars –- that is, as a named tuple in the form (Sₚ = :salt_name, θ = :potential_temp_name) where :potential_temp_name and :salt_name are the name of the potential temperature and salinity in the RasterStack.

OceanRasterConversions.get_ρ — Function

function get_ρ(Sₐ::Raster, Θ::Raster, p::Raster)
function get_ρ(stack::RasterStack, var_names::NamedTuple)
function get_ρ(series::RasterStack, var_names::NamedTuple)

Compute in-situ density, ρ, using gsw_rho from GibbsSeaWater.jl. This computation depends on absolute salinity (Sₐ), conservative temperature (Θ) and pressure (p). To compute ρ from a RasterStack or RasterSeries the variable names must be passed into the function as a NamedTuple in the form (Sₐ = :salt_var, Θ = :temp_var, p = :pressure_var). The returned Raster will have the same dimensions as Rasterstack that is passed in.

OceanRasterConversions.get_σₚ — Function

function get_σₚ(Sₐ::Raster, Θ::Raster, p::Float64)
function get_σₚ(stack::RasterStack, var_names::NamedTuple)
function get_σₚ(series::RasterStack, var_names::NamedTuple)

Compute potential density at reference pressure p, σₚ, using gsw_rho from GibbsSeaWater.jl. This computation depends on absolute salinity (Sₐ), conservative temperature (Θ) and a user entered reference pressure (p). Compute and return the potential density σₚ at reference pressure p from a RasterStack or RasterSeries. This computation depends on absolute salinity Sₐ, conservative temperature Θ and a reference pressure p. The variable names must be passed into the function as a NamedTuple in the form (Sₐ = :salt_var, Θ = :temp_var, p = ref_pressure). Note p in this case is a number. The returned Raster will have the same dimensions as Rasterstack that is passed in.