Function index

OceanVariableConversions

OceanVariableConversions exported functions

OceanRasterConversions.OceanVariableConversions — Module

Module to convert variables depth, practical salinity and potential temperature to the TEOS-10 standard variables pressure, absolute salinity and conservative temperature (respectively) from a Raster, RasterStack or RasterSeries. A few chosen seawater variables can then be computed from these state variables.

OceanRasterConversions.OceanVariableConversions.Sₚ_to_Sₐ — Method

function Sₚ_to_Sₐ(Sₚ::Raster)
function Sₚ_to_Sₐ(stack::RasterStack, Sₚ::Symbol)
function Sₚ_to_Sₐ(series::RasterSeries, Sₚ::Symbol)

Convert a Raster of practical salinity (Sₚ) to absolute salinity (Sₐ) using gsw_sa_from_sp from GibbsSeaWater.jl. This conversion depends on pressure. If converting from a RasterStack or RasterSeries, the symbol for the practical salinity in the RasterStack/Series must be passed in as a Symbol –- that is if the variable name is SALT the RasterStack/Series, the Symbol :SALT must be passed in.

OceanRasterConversions.OceanVariableConversions.convert_ocean_vars — Method

function convert_ocean_vars(raster::RasterStack, var_names::NamedTuple;
                            ref_pressure = nothing,
                            with_α = false,
                            with_β = false)
function convert_ocean_vars(raster::Rasterseries, var_names::NamedTuple;
                            ref_pressure = nothing,
                            with_α = false,
                            with_β = false)

Convert ocean variables depth, practical salinity and potential temperature to pressure, absolute salinity and conservative temperature. All conversions are done using the julia implementation of TEOS-10 GibbsSeaWater.jl. A new Raster is returned that contains the variables pressure, absolute salinity, conservative temperature and density (either in-situ or referenced to a user defined reference pressure). As pressure depends on latitude and depth, it is added as a new variable –- that is, each longitude, latitude, depth and time have a variable for pressure. A density variable is also computed which, by default, is in-situ density. Potential density at a reference pressure can be computed instead by passing a the keyword argument ref_pressure. Optional keyword arguments with_α and with_β allow the thermal expansion and haline contraction coefficients (respectively) to be computed and added to the returned RasterStack/Series.

The name of the variables for potential temperature and practical salinity must be passed in as a NamedTuple of the form (Sₚ = :salt_name, θ = :potential_temp_name) where :potential_temp_name and :salt_name are the name of the potential temperature and practical salinity in the Raster.

OceanRasterConversions.OceanVariableConversions.depth_to_pressure — Method

function depth_to_pressure(raster::Raster)
function depth_to_pressure(stack::RasterStack)

Convert the depth dimension (Z) to pressure using gsw_p_from_z from GibbsSeaWater.jl. Note that pressure depends on depth and latitude so the returned pressure is stored as a variable in the resulting Raster rather than replacing the vertical depth dimension.

OceanRasterConversions.OceanVariableConversions.get_α — Method

function get_α(Sₐ::Raster, Θ::Raster, p::Raster)
function get_α(stack::RasterStack, var_names::NamedTuple)
function get_α(series::RasterSeries, var_names::NamedTuple)

Compute the thermal exapnsion coefficient, α, using gsw_alpha from GibbsSeaWater.jl. To compute α from a RasterStack or RasterSeries the variable names must be passed into the function as a NamedTuple in the form (Sₐ = :salt_var, Θ = :temp_var, p = :pressure_var). The returned Raster will have the same dimensions as Rasterstack that is passed in.

OceanRasterConversions.OceanVariableConversions.get_β — Method

function get_β(Sₐ::Raster, Θ::Raster, p::Raster)
function get_β(stack::RasterStack, var_names::NamedTuple)
function get_β(series::RasterSeries, var_names::NamedTuple)

Compute the haline contraction coefficient, β, using gsw_beta from GibbsSeaWater.jl. To compute β from a RasterStack or RasterSeries the variable names must be passed into the function as a NamedTuple in the form (Sₐ = :salt_var, Θ = :temp_var, p = :pressure_var). The returned Raster will have the same dimensions as Rasterstack that is passed in.

OceanRasterConversions.OceanVariableConversions.get_ρ — Method

function get_ρ(Sₐ::Raster, Θ::Raster, p::Raster)
function get_ρ(stack::RasterStack, var_names::NamedTuple)
function get_ρ(series::RasterStack, var_names::NamedTuple)

Compute in-situ density, ρ, using gsw_rho from GibbsSeaWater.jl. This computation depends on absolute salinity (Sₐ), conservative temperature (Θ) and pressure (p). To compute ρ from a RasterStack or RasterSeries the variable names must be passed into the function as a NamedTuple in the form (Sₐ = :salt_var, Θ = :temp_var, p = :pressure_var). The returned Raster will have the same dimensions as Rasterstack that is passed in.

OceanRasterConversions.OceanVariableConversions.get_σₚ — Method

function get_σₚ(Sₐ::Raster, Θ::Raster, p::Number)
function get_σₚ(stack::RasterStack, var_names::NamedTuple)
function get_σₚ(series::RasterStack, var_names::NamedTuple)

Compute potential density at reference pressure p, σₚ, using gsw_rho from GibbsSeaWater.jl. This computation depends on absolute salinity (Sₐ), conservative temperature (Θ) and a user entered reference pressure (p). Compute and return the potential density σₚ at reference pressure p from a RasterStack or RasterSeries. This computation depends on absolute salinity Sₐ, conservative temperature Θ and a reference pressure p. The variable names must be passed into the function as a NamedTuple in the form (Sₐ = :salt_var, Θ = :temp_var, p = ref_pressure). Note p in this case is a number. The returned Raster will have the same dimensions as Rasterstack that is passed in.

OceanRasterConversions.OceanVariableConversions.θ_to_Θ — Method

function θ_to_Θ(θ::Raster, Sₐ::Raster)
function θ_to_Θ(stack::RasterStack, var_names::NamedTuple)
function θ_to_Θ(series::RasterSeries, var_names::NamedTuple)

Convert a Raster of potential temperature (θ) to conservative temperature (Θ) using gsw_ct_from_pt from GibbsSeaWater.jl. This conversion depends on absolute salinity. If converting from a from a RasterStack or a RasterSeries, the var_names must be passed in as for convert_ocean_vars –- that is, as a named tuple in the form (Sₚ = :salt_name, θ = :potential_temp_name) where :potential_temp_name and :salt_name are the name of the potential temperature and salinity in the RasterStack.

OceanVariableConversions private functions

OceanRasterConversions.OceanVariableConversions.get_dims — Method

function get_dims(raster::Raster)

Get the dimensions of a Raster.

RasterHistograms

RasterHistograms exported functions

OceanRasterConversions.RasterHistograms — Module

Module to fit Histograms to data that is saved as a Raster, RasterStack or RasterSeries. The fitted Histogram will be N-dimensional, where N is the number of layers (i.e. variables) to fit the Histogram to. The Histogram is fitted to the Raster data using the StatsBase.jl] empirical estimation module. For more information on how the Histogram is fitted to the data, and the arguments see the StatsBase.jl documentation.

OceanRasterConversions.RasterHistograms.RasterLayerHistogram — Type

mutable struct RasterLayerHistogram <: AbstractRasterHistogram

A RasterLayerHistogram. The struct is mutable so that the histogram field can be updated using the normalize (or otherwise) function.

layer::Symbol: The layer (variable) from the Raster

dimensions::Tuple: The dimensions of the Raster

raster_size::Tuple: The size of the Raster

histogram::StatsBase.AbstractHistogram: The 1-dimensional histogram fitted to the Raster layer data

OceanRasterConversions.RasterHistograms.RasterLayerHistogram — Method

function RasterLayerHistogram(rs::Raster; closed = :left, nbins = nothing)
function RasterLayerHistogram(rs::Raster, weights::AbstractWeights;
                              closed = :left, nbins = nothing)
function RasterLayerHistogram(rs::Raster, edges::AbstractVector; closed = :left)
function RasterLayerHistogram(rs::Raster, weights::AbstractWeights,
                              edges::AbstractVector; closed = :left)

Construct a RasterLayerHistogram from a Raster. The flattened Raster data, with the missing values removed, is passed to the fit(::Histogram) function from StatsBase.jl and a RasterLayerHistogram type is returned.

OceanRasterConversions.RasterHistograms.RasterSeriesHistogram — Type

mutable struct RasterSeriesHistogram <: AbstractRasterHistogram

A RasterSeriesHistogram. The struct is mutable so that the histogram field can be updated using the normalize (or otherwise) function.

layers::Tuple: The layers (variables) from the RasterSeries used to fit the Histogram

series_dimension::Symbol: The dimension of the RasterSeries (usually this will be time)

series_length::Int64: The length of the `RasterSeries

dimensions::Tuple: The dimensions of the elements (either a Raster or RasterStack) of the RasterSeries

raster_size::Tuple: The size of the elements (either a Raster or RasterStack) of the RasterSeries

histogram::StatsBase.Histogram: The N-dimensional Histogram fitted to the N layers from RasterSeries

OceanRasterConversions.RasterHistograms.RasterSeriesHistogram — Method

function RasterSeriesHistogram(series::RasterSeries, edges::NTuple{N, AbstractVector};
                               closed = :left)
function RasterSeriesHistogram(series::RasterSeries, weights::AbstractWeights,
                               edges::NTuple{N, AbstractVector}; closed = :left)

Construct a RasterSeriesHistogram from a RasterSeries. Note that to merge Histograms the bin edges must be the same, so for this constructor the edges must be passed in. This constructor assumes that the dimensions are the same across all RasterStacks in the RasterSeries.

OceanRasterConversions.RasterHistograms.RasterStackHistogram — Type

mutable struct RasterStackHistogram <: AbstractRasterHistogram

A RasterStackHistogram. The struct is mutable so that the histogram field can be updated using the normalize (or otherwise) function.

layers::Tuple: The layers (variables) from the RasterStack used to fit the Histogram

dimensions::Tuple: The dimensions of the RasterStack

raster_size::Tuple: The size of the RasterStack layers

histogram::StatsBase.Histogram: The N-dimensional Histogram fitted to the N layers from RasterStack

OceanRasterConversions.RasterHistograms.RasterStackHistogram — Method

function RasterStackHistogram(stack::RasterStack; closed = :left, nbins = nothing)
function RasterStackHistogram(stack::RasterStack, weights::AbstractWeights;
                              closed = :left, nbins = nothing)
function RasterStackHistogram(stack::RasterStack, edges::NTuple{N, AbstractVector};
                              closed = :left)
function RasterStackHistogram(stack::RasterStack, weights::AbstractWeights,
                              edges::NTuple{N, AbstractVector}; closed = :left)

Construct a RasterStackHistogram from a RasterStack. The resulting Histogram is N-dimensional, where N is the number of layers. The flattened Raster data for each layer, with themissing values removed, is passed to the fit(::Histogram) function from StatsBase.jl and a RasterStackHistogram type is returned.

LinearAlgebra.normalize! — Method

function normalize!(arh::AbstractRasterHistogram; mode::Symbol = :pdf)

Normalize the Histogram in the AbstractRasterHistogram according the desired mode. See the StatsBase.jl docs for information on the possible modes and how they work.

MakieCore.convert_arguments — Function

function convert_arguments(P::Type{<:AbstractPlot}, arh::AbstractRasterHistogram)

Converting method so Makie.jl can plot an AbstractRasterHistogram. Note for plotting purposes a value correspnding to zero is replaced with NaN. This is to avoid in 2D plotting many zero values and as a result not seeing the distribtution clearly (in Makie.kj heatmap the NaN value is left out in plotting).

OceanRasterConversions.RasterHistograms.area_weights — Method

function area_weights(rs::Union{Raster, RasterStack}; equator_one_degree = 111e3)

Return the Weights for a Histogram calculated from the area of each grid cell in a Raster or RasterStack. The Raster or RasterStack must first be sliced over the dimensions one wishes to look at, e.g. for area weights at sea surface the function need rs[Z(1)] to be passed in. If the original Raster only has two spatial dimensions then this step may be skipped. The keyword argument equator_one_degree is one degree at the equator in metres. The function returns a container Weights so can be passed straight into the fit(::Histogram) function.

OceanRasterConversions.RasterHistograms.volume_weights — Method

function volume_weights(rs::Union{Raster, RasterStack}; equator_one_degree = 111e3)

Return the Weights for a Histogram calculated from the volume of each grid cell in a Raster or RasterStack. The model resolution is inferred from the X and Y dimensions of the Raster or RasterStack and assumes that along the X and Y the resolution is unique (though it can be different for X and Y). The keyword argument equator_one_degree is one degree at the equator in metres. The function returns a container Weights so can be passed straight into the fit(::Histogram) function.

RasterHistograms private functions

OceanRasterConversions.RasterHistograms.AbstractRasterHistogram — Type

Abstract supertype for a RasterHistogram.

OceanRasterConversions.RasterHistograms.find_stack_non_missing — Method

function find_stack_non_missing(stack::RasterStack)

Return a Raster of type Bool that contains the intersection of the non-missing values from the layers of a RasterStack.

OceanRasterConversions.RasterHistograms.raster_zeros_to_nan — Method

function raster_zeros_to_nan(arh::AbstractRasterHistogram)

Convert the zeros (i.e. empty bins in the AbstractRasterHistogram) to NaNs for plotting in Makie.

RecipesBase.apply_recipe — Method

Conversion method for plotting in Plots.jl